"rename the file "GenerateRiverFlows.py" to 'generate_river_flows.py'"
